add add_warning to excpetions.h
